step3.py
import pandas as pd
import numpy as np
import riskfolio as rp
import matplotlib.pyplot as plt
from typing import Tuple

def calculate_returns(data: pd.DataFrame, assets: list) -> pd.DataFrame:
    """
    Calculate the returns of the given assets.
    """
    return data[assets].pct_change().dropna()

def build_portfolio(returns: pd.DataFrame) -> rp.Portfolio:
    """
    Build a portfolio object using the given returns.
    """
    return rp.Portfolio(returns=returns)

def estimate_assets_stats(portfolio: rp.Portfolio, method_mu: str = 'hist', method_cov: str = 'hist', d: float = 0.94) -> None:
    """
    Estimate asset statistics using the given methods.
    """
    portfolio.assets_stats(method_mu=method_mu, method_cov=method_cov, d=d)

def optimize_portfolio(portfolio: rp.Portfolio, model: str = 'Classic', rm: str = 'MV', obj: str = 'Sharpe', rf: float = 0, l: float = 0, hist: bool = True) -> pd.DataFrame:
    """
    Optimize the portfolio using the given parameters.
    """
    return portfolio.optimization(model=model, rm=rm, obj=obj, rf=rf, l=l, hist=hist)

def plot_pie(weights: pd.DataFrame) -> None:
    """
    Plot the composition of the portfolio as a pie chart.
    """
    rp.plot_pie(w=weights, title='Sharpe Mean Variance', others=0.05, nrow=25, cmap="tab20", height=6, width=10, ax=None)
    plt.show()

def efficient_frontier(portfolio: rp.Portfolio, model: str = 'Classic', rm: str = 'MV', points: int = 50, rf: float = 0, hist: bool = True) -> pd.DataFrame:
    """
    Calculate the efficient frontier of the portfolio.
    """
    return portfolio.efficient_frontier(model=model, rm=rm, points=points, rf=rf, hist=hist)

def plot_frontier(portfolio: rp.Portfolio, frontier: pd.DataFrame, rm: str = 'MV', rf: float = 0, label: str = 'Max Risk Adjusted Return Portfolio') -> None:
    """
    Plot the efficient frontier of the portfolio.
    """
    ax = rp.plot_frontier(w_frontier=frontier, mu=portfolio.mu, cov=portfolio.cov, returns=portfolio.returns, rm=rm, rf=rf, alpha=0.05, cmap='viridis', w=portfolio.w, label=label, marker='*', s=16, c='r', height=6, width=10, ax=None)
    plt.show()

def plot_frontier_area(frontier: pd.DataFrame) -> None:
    """
    Plot the efficient frontier composition as an area chart.
    """
    rp.plot_frontier_area(w_frontier=frontier, cmap="tab20", height=6, width=10, ax=None)
    plt.show()

if __name__ == '__main__':
    data = ... # Load data here
    assets = ... # Define assets here
    time_horizon = ... # Define time horizon here
    metric = ... # Define metric here

    returns = calculate_returns(data, assets)
    portfolio = build_portfolio(returns)
    estimate_assets_stats(portfolio)
    weights = optimize_portfolio(portfolio,
